# openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/api/balance/{user_id}": {"get": {"summary": "Status", "operationId": "status_api_balance__user_id__get", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/inc": {"put": {"summary": "Inc", "operationId": "inc_api_inc_put", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateData"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/reset": {"put": {"summary": "Reset", "operationId": "reset_api_reset_put", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateData"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/user/login/{user_id}": {"get": {"summary": "Login", "operationId": "login_api_user_login__user_id__get", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/user/sync": {"post": {"summary": "Sync Data", "operationId": "sync_data_api_user_sync_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResponseData"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ResponseData": {"properties": {"user_id": {"type": "integer", "title": "User Id"}, "session_id": {"type": "integer", "title": "Session Id"}, "balance": {"type": "integer", "title": "Balance"}, "status": {"$ref": "#/components/schemas/Status"}}, "type": "object", "title": "ResponseData"}, "Status": {"type": "string", "enum": ["Available", "In-use"], "title": "Status"}, "UpdateData": {"properties": {"password": {"type": "string", "title": "Password"}, "user_id": {"type": "integer", "title": "User Id"}, "diff": {"type": "integer", "title": "Diff"}}, "type": "object", "title": "UpdateData"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}